requirements.txt
db-sqlite3==0.0.1
fastapi==0.115.12
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
pydantic-settings==2.9.1
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
sqlmodel==0.0.24
